server/requirements.txt
openai
langchain
langchain_community
langchain_openai
langsmith
python-dotenv
flask